CSP Hashes for Embedded `<script>` & `<style>`
Calculate hashes for embedded `<script>` and `<style>` tags, for use
with the `script-src` and `style-src` directives of the
`Content-Security-Policy` HTTP header, to mitigate cross-site scripting
(XSS) attacks.

This belongs here, and not in `JS::Packer` and `CSS::Packer`, because
hashes are only relevant for embedded scripts and stylesheets. Scripts
and stylesheets served separately can be whitelisted by path, which is
much simpler and more performant than hashing.